requirements.txt
astroid==2.4.2
attrs==20.2.0
colorama==0.4.3
coverage==5.3
docopt==0.6.2
iniconfig==1.0.1
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pathtools==0.1.2
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.2
pytest-describe==1.0.0
pytest-testmon==1.0.3
pytest-watch==4.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
toml==0.10.1
watchdog==0.10.3
wrapt==1.12.1
